Move AppDatabase location
