+++ app/src/main/java/com/example/volcanoseason3/data/database/AppDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.volcanoseason3.data.forecastLinks
+package com.example.volcanoseason3.data.database
 
 import android.content.Context
 import androidx.room.Database
@@ -6,6 +6,8 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
+import com.example.volcanoseason3.data.forecastLinks.ForecastLink
+import com.example.volcanoseason3.data.forecastLinks.ForecastLinkDao
 
 const val DATABASE_NAME = "forecast-links-db"
